7cfd942 Merge branch 'main' into dev  @SupenBysz

ff17a9d feat(sys_auth): 优化用户注册时的邀约码验证逻辑  @SupenBysz
876796b feat(sys_user): 在用户创建逻辑中添加邀约码字段  @SupenBysz
6891375 feat(sys_member_level): 新增根据用户ID获取会员等级列表功能  @SupenBysz
de514b0 pref(sys_rules): 优化邀约码检查逻辑  @SupenBysz
62b1cbe feat(sys_permission): 添加API权限白名单功能  @SupenBysz
ffee668 feat(sys_mail, sys_sms): 在验证逻辑中添加开发模式支持  @SupenBysz
d4da14b fix(sys_audit): 修改 QueryAuditList 方法参数名称  @SupenBysz
891b201 feat(invite): 重构邀约系统功能并新增邀请码字段  @SupenBysz
2c9bd36 feat(会员等级): 添加level字段表示权益级别  @SupenBysz
68e628c fix: 修正邀请码字段名拼写错误  @SupenBysz
d65a148 refactor(sys_audit): 更新审核状态枚举值及注释  @SupenBysz
c2d357b docs(sys_file): 更新AllowAnonymous字段的注释说明  @SupenBysz
8de8afc fix: 修正字段名拼写错误，将InviceCode改为InviteCode  @SupenBysz
1cf4428 fix: 修正 BaiduSdkConfAccessToken 结构体中 refreshToken 字段命名  @SupenBysz
a38d62d fix: 将评分字段类型从float64改为int  @SupenBysz
9ec4984 feat(invite): 添加设置邀请者单位标识前缀功能并优化创建逻辑  @SupenBysz
9aa8be1 fix: 修正更新邀请者单位标识前缀的条件字段  @SupenBysz
c8dfcba feat: 添加启用邀请注册功能的配置字段 @SupenBysz
905b072 refactor(会员等级): 优化会员等级用户相关接口返回类型及逻辑 @SupenBysz

+++ sys_model/sys_entity/sys_comment.go
@@ -17,7 +17,7 @@ type SysComment struct {
 	Body          string      `json:"body"          orm:"body"            description:"图文评论"`
 	MediaIds      string      `json:"mediaIds"      orm:"media_ids"       description:"媒体资源：图文、视频等"`
 	ReplyId       int64       `json:"replyId"       orm:"reply_id"        description:"评论回复信息ID，即关联父级评论ID"`
-	Score         float64     `json:"score"         orm:"score"           description:"评分0-5，间隔0.1"`
+	Score         int         `json:"score"         orm:"score"           description:"评分0-5，间隔0.1"`
 	CreatedAt     *gtime.Time `json:"createdAt"     orm:"created_at"      description:"评论发表时间"`
 	UnionId       int64       `json:"unionId"       orm:"union_id"        description:"关联业务ID"`
 }
